Updated to v 1.6.2 Fixed loss of stereointegrity issue Fixed canonicalisation issue resulting from above fix Fixed graceless failures for some salted forms and apparent successful failures of other salted forms

+++ com.vernalis.knime.chem.mmp/src/com/vernalis/knime/mmp/FragmentValue2.java
@@ -287,8 +287,8 @@ public class FragmentValue2 implements Comparable<FragmentValue2> {
 		for (Entry<Integer, Integer> ent : apLookup.entrySet()) {
 			SMILES = SMILES.replace("[" + ent.getValue() + "*]", "[" + ent.getKey() + "*]");
 		}
-		canonicalizeBondOnlyValue();
-		// canonicalize();
+		// canonicalizeBondOnlyValue();
+		canonicalize();
 	}
 
 	/**
@@ -296,9 +296,13 @@ public class FragmentValue2 implements Comparable<FragmentValue2> {
 	 * {@link ROMol} object.
 	 */
 	public void canonicalize() {
-		ROMol mol = RWMol.MolFromSmiles(SMILES, 0, false);
-		SMILES = mol.MolToSmiles(true);
-		mol.delete();
+		if (SMILES.matches("\\[\\d+\\*\\](-)?\\[\\d+\\*\\]")) {
+			canonicalizeBondOnlyValue();
+		} else {
+			ROMol mol = RWMol.MolFromSmiles(SMILES, 0, false);
+			SMILES = mol.MolToSmiles(true);
+			mol.delete();
+		}
 	}
 
 	/**
